requirements.txt
# Requirements for Cisco PnP Automation Project

# Core dependencies
requests>=2.31.0
PyYAML>=6.0.1
Jinja2>=3.1.2

# Optional but recommended
urllib3>=1.26.16
certifi>=2023.7.22

# Development and testing
pytest>=7.4.0
pytest-mock>=3.11.1

# Logging and utilities  
colorlog>=6.7.0
python-dotenv>=1.0.0

# Documentation
Sphinx>=7.1.0
sphinx-rtd-theme>=1.3.0

# For additional network automation (optional)
netmiko>=4.2.0
napalm>=4.1.0
paramiko>=3.2.0

# JSON/YAML processing
jsonschema>=4.18.0
ruamel.yaml>=0.17.32